app: move the usecase layer out of core
