+++ app/delete.go
