+++ cmd/list/list.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/zilliztech/milvus-backup/cmd/flags"
 	"github.com/zilliztech/milvus-backup/cmd/root"
-	"github.com/zilliztech/milvus-backup/core/app"
+	"github.com/zilliztech/milvus-backup/app"
 	v2 "github.com/zilliztech/milvus-backup/internal/cfg/v2"
 )
